part

+++ src/main/java/com/myapp/Controller/user/UserController.java
@@ -1,9 +1,6 @@
 package com.myapp.Controller.user;
 
-import com.myapp.DB.ProducerRepository;
-import com.myapp.DB.ProductRepository;
-import com.myapp.DB.User;
-import com.myapp.DB.UserRepository;
+import com.myapp.DB.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Controller;
@@ -13,6 +10,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
+
+import java.util.LinkedList;
+import java.util.List;
 
 /**
  * Created by Bos on 2017-01-08.
@@ -69,24 +69,31 @@ public class UserController {
     ) {
         ModelAndView modelAndView = null;
         if( action.equals("editItem") ){
-            //handle save
+            modelAndView = new ModelAndView("home/editUserItems", "command", new commandEditItems());
+            modelAndView.addObject("user", userRepository.findOne(command.getId()));
+            modelAndView.addObject("availItem", productRepository.findByIsBorrowed(new Long(0)));
+
+
         }
         else if( action.equals("editUser") ){
-
-
             modelAndView = new ModelAndView("home/editUser","command", new command());
             modelAndView.addObject("user", userRepository.findOne(command.getId()));
         }
-
-
-
         else if( action.equals("delUser") ){
+            if(userRepository.findOne(command.getId()).getProducts().isEmpty()) {
+                userRepository.delete(command.getId());
+                modelAndView = new ModelAndView("home/result");
+                modelAndView.addObject("resultCause", 1);
+                modelAndView.addObject("msg", "You successfully deleted user!");
+                modelAndView.addObject("back", "/home/users");
+            } else {
 
-            userRepository.delete(command.getId());
-            modelAndView = new ModelAndView("home/result");
-            modelAndView.addObject("resultCause", 1);
-            modelAndView.addObject("msg", "You successfully deleted user!");
-            modelAndView.addObject("back", "/home/users");
+                modelAndView = new ModelAndView("home/result");
+                modelAndView.addObject("resultCause", 0);
+                modelAndView.addObject("msg", "No! This user have some products!!");
+                modelAndView.addObject("back", "/home/users");
+            }
+
 
         }
         else
@@ -106,12 +113,46 @@ public class UserController {
     public ModelAndView editConfirm(@ModelAttribute("SpringWeb")command command,
                                     ModelMap model) {
 
-        User newUser = new User(command.getId(),command.getFirstName(),command.getLastName());
+        User newUser = userRepository.findOne(command.getId());
+        newUser.setFirstName(command.getFirstName());
+        newUser.setLastName(command.getLastName());
         userRepository.save(newUser);
 
         ModelAndView modelAndView = new ModelAndView("home/result");
         modelAndView.addObject("resultCause", 1);
         modelAndView.addObject("msg", "You successfully edited user!");
+        modelAndView.addObject("back", "/home/users");
+
+        return modelAndView;
+    }
+
+
+    @RequestMapping(value = "/home/editConfirmItem", method = RequestMethod.POST)
+    public ModelAndView editConfirm(@ModelAttribute("SpringWeb")commandEditItems command,
+                                    ModelMap model) {
+
+
+
+        User newUser = userRepository.findOne(command.getUserId());
+
+        for (Product pord:newUser.getProducts()) {
+            pord.setIsBorrowed(new Long(0));
+            productRepository.save(pord);
+        }
+
+        List<Product> list = new LinkedList<Product>();
+        for (Long id:command.getUserItems()) {
+            Product item = productRepository.findOne(id);
+            list.add(item);
+            item.setIsBorrowed(new Long(1));
+            productRepository.save(item);
+        }
+        newUser.setProducts(list);
+        userRepository.save(newUser);
+
+        ModelAndView modelAndView = new ModelAndView("home/result");
+        modelAndView.addObject("resultCause", 1);
+        modelAndView.addObject("msg", "You successfully edited user items!");
         modelAndView.addObject("back", "/home/users");
 
         return modelAndView;
